requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
asteroid-filterbanks==0.4.0
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blessed==1.20.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmaes==0.9.1
colorama==0.4.6
colorlog==6.7.0
contourpy==1.0.7
croniter==1.3.15
cycler==0.11.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.3.0
dnspython==2.3.0
docopt==0.6.2
einops==0.6.1
fastapi==0.88.0
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
h11==0.14.0
huggingface-hub==0.15.1
HyperPyYAML==1.2.1
idna==3.4
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
julius==0.2.7
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.10.0.post2
lightning==2.0.2
lightning-cloud==0.5.36
lightning-utilities==0.8.0
llvmlite==0.40.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
motor==3.1.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
numba==0.57.0
numpy==1.24.3
odmantic==0.9.2
omegaconf==2.3.0
optuna==3.2.0
ordered-set==4.1.0
packaging==23.1
pandas==2.0.2
passlib==1.7.4
Pillow==9.5.0
pooch==1.6.0
primePy==1.3
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.7.0
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytorch-metric-learning==2.1.2
pytz==2023.3
PyYAML==6.0
readchar==4.0.5
regex==2023.6.3
requests==2.29.0
rich==13.4.1
rsa==4.9
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
semver==3.0.0
sentencepiece==0.1.99
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
speechbrain==0.5.14
SQLAlchemy==2.0.15
starlette==0.22.0
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
wcwidth==0.2.6
websocket-client==1.5.2
websockets==11.0.3
yarl==1.9.2
